+++ src/qwip/processing/processors.py
@@ -829,46 +829,50 @@ class Labeled(GenericDataProcessor):
     def run(
         self,
         result: M,
-        labels: pd.Index | None = None,
+        label: pd.Index | None = None,
         exe: QuantumExecutable | None = None,
         **kwargs,
     ) -> M:
-        if labels is None and (exe is None or exe.seq is None):
+        if label is None and (exe is None or exe.seq is None):
             return result
-
-        result = attrs.evolve(result, data=result.data.copy())
-        old_idx = result.d.index
-
-        if labels is None:
+        elif label is None:
             seq = exe.sequence
-            labels = pd.DataFrame(
-                it.product(
-                    *(
-                        seq.labels.get(n, np.arange(seq.shape[i]))
-                        for i, n in enumerate(seq.names)
-                    )
-                ),
-                columns=[
-                    name or f"{self.level}{i}" for i, name in enumerate(seq.names)
-                ],
-                index=old_idx.levels[old_idx.names.index(self.level)],
+            label = seq.flatten().labels[0]
+        elif not isinstance(label, pd.Index):
+            raise ValueError(
+                f"Explicit labels must be a pandas index, got {type(label)}."
             )
 
-        broadcasted = labels.loc[old_idx.get_level_values(self.level)]
+        old_idx = result.data.index
+        level_to_relabel = old_idx.levels[old_idx.names.index(self.level)]
+        if len(label) != len(level_to_relabel):
+            raise ValueError(
+                f"Number of labels {len(label)} does not match length "
+                f"{len(level_to_relabel)} of level {self.level}."
+            )
 
-        idx_vals = []
-        idx_names = []
-        for name in old_idx.names:
-            if name == self.level:
-                for c in labels.columns:
-                    idx_vals.append(broadcasted[c].values)
-                    idx_names.append(c)
+        result = attrs.evolve(result, data=result.data.copy())
+
+        names = []
+        levels = []
+        codes = []
+        for name, level, code in zip(old_idx.names, old_idx.levels, old_idx.codes):
+            if name != self.level:
+                names.append(name)
+                levels.append(level)
+                codes.append(code)
+            elif isinstance(label, pd.MultiIndex):
+                names.extend(label.names)
+                levels.extend(label.levels)
+
+                for label_code in label.codes:
+                    codes.append(label_code[code])
             else:
-                level = old_idx.get_level_values(name)
-                idx_vals.append(level.values)
-                idx_names.append(level.name)
+                names.append(label.name)
+                levels.append(label)
+                codes.append(code)
 
-        result.data.index = pd.MultiIndex.from_arrays(idx_vals, names=idx_names)
+        result.data.index = pd.MultiIndex(levels=levels, codes=codes, names=names)
 
         return result